Replaced previous OpenSSL integration with our Parraid.OpenSSL Nuget package
Unfortunately, many (all?) of the .vcxproj project files of pjproject are in
the same directory, so if you add a package to one project, the package
manager thinks it was added to many other projects too, because it sees
the package.config in the same directory as those other project files, too.
This is a minor nuisance because many projects don't need OpenSSL.

Also, this caused problems when I was preparing this change: Projects
that didn't have OpenSSL showed up as already having it, and therefore
didn't allow installing. So I had to pretend-uninstall it and the reinstall
it, which of course also uninstalled and reinstalled it in other projects.

Furthermore, when I updated the package in the Github repo, the
package manager of course suggested that I upgrade all projects that
it thinks have it installed, including those which really don't. This may
get somewhat messy in the future but if any projects end up getting
the package installed even if they don't need it, it's not a big deal.

The samples project unfortunately doesn't build in its current state.
It's based on a makefile and unfortunately the library location for
libcrypto.lib and libssl.lib is not passed to nmake. I could probably
fix it but it's not that important.

Some C and C++ files still generate warnings. I looked at them and
started fixing them but decided they all looked innocent except for
one in ssl_sock.ossl.c. So I fixed that one but not the other ones.
I figure once we merge the latest from upstream we can see if the
problems are still there and re-evaluate.